"Refactor code for clarity and expand calculation display
Brought import statements of same library under one umbrella and rearranged them for better readability in 'main.py'. Streamlined the creation of App class with well-structured calls and functional adjustments. Additionally, some adjustments on input panel were carried out.

In 'home.py', comprehensively expanded the DisplayResults class for a detailed output. Outputs are now better formatted, making them easier to understand. Also, improved the usage of the labels, values and color mapping in pie chart display, providing a better visual representation of the financial breakdown.

Revamped the OptimizeIncome class to perform optimization in an enhanced and more readable way. Brought in significant definitions for a clearer initialization of this class."