Refactor code and update default input values
This commit encompasses a reshuffling and simplification of the code structure within home.py. The optimization parameters and results display have been separated into its own function which enhances readability and code organisation. Default input values in the GUI have been updated to make them more realistic and user-friendly. This refactor has no impact on the functionality of the code but improves the user experience and maintainability of the code base.